{"schema":"decodex/commit/1","summary":"Split app server client endpoint modules","authority":"manual"}
